rename legacy implementation
