+++ plugin/evm/order_api.go
@@ -72,7 +72,7 @@ func (api *OrderAPI) PlaceSignedOrders(ctx context.Context, input string) (Place
 		ordersToGossip = append(ordersToGossip, order)
 	}
 
-	api.vm.legacyGossiper.GossipSignedOrders(ordersToGossip)
+	api.vm.orderGossiper.GossipSignedOrders(ordersToGossip)
 
 	return PlaceSignedOrdersResponse{Orders: response}, nil
 }
